Combine replykeyboard stuff into single file
